infra: reorganize the `RELEASE_CHECKLIST` to put version changes together.
We previously created a pattern where some of the manual version changes
that needed to be made were appearing in the "Prepare a Changelog (Full
release only)" section and some were in the "Start a release PR" section.

This wasn't the right organization and I found this to be counterintuiative
since many of the version bumps were actually important to make even for
non-full-releases and further challenging because they weren't about the
changelog but rather version bumps which were also happening in the other
section.

This puts all the (too many) steps in one ordered list.