extract and group various configuration elements from server
This introduces three new top-level configuration groupings:

 - Listeners (currently only 1 for data, but will add admin support)
 - Paths (health, graphql)
 - GraphqQL (introspection, experimental_defer, experimental_recursion)

and retains the existing Server configuration.
 - Server (cors, sandbox)

The purpose of the grouping is to be functional in nature. Clearly,
plugins don't fit in with this, but that's one for the future...

Various configuration options are extracted from Server into these new
groupings. The purpose of the new groups

Some renaming is done as well, e.g.: landing_page -> sandbox, endpoint
-> graphql

The docs are mainly modified, but I would like to re-structure them a
bit to reflect this grouping.

I'm aware that this goes way beyond the remit of 1606, but I figured I
may as well do some tidying up now, since the current situation is far
from ideal.

fixes: #1606