Include the `v` in the release artifact tarball.
The fact that we introduced such subtle changes to code that was proven to
both work and do exactly what we wanted has been really frustrating to chase
down.  Cargo culting is one thing, but to cargo cult and just make small
nuanced and arguably unnecessary changes that cannot be easily reconciled by
humans or tooling is extra cognitive overhead that we can do without.

There's certainly the argument that we could have written a library, but the
rule of threes is still a pretty reasonable thing to do before creating
abstractions.   In the general sense, it should be trivial to let two copies
of a similar script exist and stay relatively aligned when their purposes
are identical.