add support for setting vary header in plugins
This changes the behaviour of the CorsLayer so that a default Vary
header is not set.

This means that a value can now be set from a plugin. When processing
the response (after all plugin processing is completed), if there is no
Vary header, then insert one with a value of "origin".

fixes: #1297