field_level_instrumentation -> field_level_instrumentation_sampler. Sample rate is defaulted to 0.01.
Both rates are an effective ratio of requests, but the field level sampling rate must be lower than the global sampling rate.

For example:
if global rate is 0.5 and field rate is also 0.5 100% of sampled traces will have field level instrumentation.
if global rate is 0.5 and field rate is also 0.25 50% of traces will have field level instrumentation.