infra: use gen2 Mac machine resource class on CircleCI + bump cache key
It's 75 credits instead of 50, but the newer generation processor may make a
huge difference in our build times, which are often getting to 20 minutes.
This can be seen in the [Product] documentation and the [Docs].

This also bumps the cache key because I don't trust the cache from the old
Mac architecture to build cleanly.

[Docs]: https://circleci.com/docs/configuration-reference#macos-execution-environment
[Product]: https://circleci.com/product/features/resource-classes/#macos