Leave out the target from log messages under default ("info") log level
(I'm not claiming this is the best implementation.  Just offering this
commit for the basis of discussion.)

This disables the (arguably, more verbose) inclusion of the internal Apollo
Router (Rust) module names (referred to as the "target" in some `env_logger`
terms) in the default log-level of "info".  They are still included in any
more verbose logging level (like, `debug` or `trace`).

The motivating reason for this ergonmics and first-run experience: the
inclusion of these module names ends up taking up a substantial portion of
each line.

Having the ability to tweak the log levels of each module is certainly a
very nice configuration surface to offer (and have access to), but it might
be an okay suggestion for someone who desires more specific log behavior to
first ask the Router for more detailed module information by using a more
verbose level.

[`env_logger`](https://docs.rs/env_logger/0.9.0/env_logger/)