replace startup with ctor for plugin registration
I've implemented the change so that we declare a function with a `paste`
generated name for each plugin. This means that we now have restrictions
on the contents of the `group` and the `name` parameters, since both of
these parameters must now only contain valid components of a function
name. That explains why I had to change "apollo.router" to
"apollo_router" for the group names in the tests.

fixes: #1703