WIP: attempt at fixing hasNext
we previously returned an empty graphql response at the end of the
response stream to set the `hasNext` field to false, to indicate that no
more responses will come.

That empty response is causing issues in some clients, so 24a00e66 was a
fix to set the `hasNext` on a deferred response from inside query
planner execution, but it does not account for parallel deferred
response executions, so one response might come with `hasNext` to false
then get another one.

This commit attempts another solution, where we go through an
intermediate task that checks if the response stream is closed (it is a
channel, so it implements `FusedStream`). Unfortuantely, right now it
fals to recognize when the stream is closed