Remove patched async-compression, disable compression for multipart responses instead
This is a different work-around for https://github.com/apollographql/router/issues/1572.

The previous work-around required a `[patch.crates-io]` section in the root `Cargo.toml`.
When the Router is used as a dependency, that root manifest is not ours
so downstream users would need to apply the same patch to get the work-around.

This came up in the context of publish the Router to crates.io
https://github.com/apollographql/router/issues/491
but applies the same when the `apollo-router` crate is used as a git dependency.