Update development guide to include info re: use of git submodules (#1752)
Add "getting started" step re: git submodules

This PR adds clarity to the Getting Started section of the Apollo Router
development guide with respect to the use of git submodules when running
the `federation-demo` for testing purposes. As it stands, there is no
mention of our use of git submodules in the guide, and running through
the process as described fails at the container build stage due to
missing files/context. While a little investigative work makes the
reason for the above occurrence clear, including this information up
front makes for a better development experience.