fix dev mode with configuration validation (#1813)
close #1811 

I changed the way we watched file changes. When we launch the file
watching it automatically triggers a first event, so before we were
reading the configuration twice. Now only once.

I had to introduce `serial_test` crate to force 2 tests for dev mode to
be ran in serial, because it needs to check the environment variable.
It's a temporary fix to let me test it and be ready for the release. But
I will find another solution in the future

Signed-off-by: Benjamin Coenen <5719034+bnjjj@users.noreply.github.com>