Removed `Request::from_bytes()` from public API
This removes `Request::from_bytes()` from the public API.  We were no
longer using it and we don't expect anyone external to have been relying on
it.

This was discovered this function during an exercise of documenting our
entire public API.  We considered keeping it briefly, but it doesn't
necessarily meet our requirements for shipping it in the public API.  It's
internal usage was removed in
[`d147f97d`](https://github.com/apollographql/router/commit/d147f97d) as part
of [PR #429](https://github.com/apollographql/router/pull/429).

We can consider re-introducing this in the future (it even has a matching
`Response::from_bytes()` which it composes against nicely!), but it seemed
worth remove it for the time-being until the use-cases are here.  (We can
already see the use-cases, but we want to actually design them first.)

Closes https://github.com/apollographql/router/issues/1855