prototyp which redacts apollo_private fields
It's not complete, but it illustrates a mechanism for solving the
problem (I think).

The JSON code is janky (but so is the code in the tracing_subscriber
crate which I've mainly just lifted here).

The plain text stuff is very simple, but I think we need a Display trait
version rather than the Debug trait stuff I've implemented. That
shouldn't be a lot of work...