Add a rust-version field in apollo-router/Cargo.toml
This field states what we want our minimum supported Rust version to be:

https://doc.rust-lang.org/cargo/reference/manifest.html#the-rust-version-field
> If the currently selected version of the Rust compiler is older than
the stated version, cargo will exit with an error, telling the user what
version is required.

This profiles a better error message than random compilation errors.

1.63.0 is our de-facto MSRV at the moment as it is what we use on CI.
I’ll file a separate issue to discuss how we want to evolve it.