fix(helm): broken prometheus annotation (#1883)
Fixes https://github.com/apollographql/router/issues/1881

I've gone back and forth a couple times on defaulting the `prometheus`
values because I caught a case where if the listen address is
unspecified but others in the object are, `helm template` fails. I'm
sure there could be some templating magic to be done but I believe it
should be avoided in favor of keeping things more robust via default
values. I'm not sure it's worth the time to drill down into the cause of
the problem when the values can just be defaulted in the chart.

A bit of scope creep here but also updates the listen default for the
healthcheck to `0.0.0.0` so it works within k8s and adds capability to
expose it so external services can use it.