Remove note about Apollo Server docs from homepage
The 'homepage' option displays the studio landing page
(https://github.com/apollographql/studio-landing-page/blob/main/src/App.tsx).
If you don't pass `footer: false`, it adds a footer that says "This page
can be customized or hidden. Learn more in the Apollo Server Docs" which
is inappropriate for Router (because it is not Apollo Server, and
because the page can't be customized other than by replacing it with a
different page.)