rework our loggers to ignore spans completely in log output (#1896)
The router creates three flavour of subscribers:
 - Console
 - Text
 - JSON

The console logger is unchanged by this work.

The JSON logger is modified so that all span fields are completely
absent from the log. This is done by making use of the
```
                        .with_current_span(false)
                        .with_span_list(false)
```
builder configuration options when the JSON logger is constructed.

Unfortunately, the Full text formatter in the tracing_subscriber crate
doesn't have equivalent builder options, so we build our own log
formatter (`TextFormatter`) using similar logic to the tracing crate but
not processing span content. It's a subset which avoids feature
configuration and makes opinionated assumptions about the output format
and doesn't provide additional configuration.

The nice thing about this change is that we can remove our hastily added
`RouterJsonFormatter` and the log text formatter doesn't have the
annoying spaces which the previous approach possessed. The downsides
are:

 - do we definitely want to remove all span content from our log files?
- are we likely to want to re-configure our text log output and require
more changes?
 - we've added a new crate (ansi-term) for ANSI control code processing

It's a neater solution than our current approach and probably is a step
in the right direction whilst we wait to see if we can get changes added
upstream to tracing to support field filtering.

(effectively, ditching all span content is equivalent to having a field
filter which removes all spans.)

There's a [long standing
issue](https://github.com/tokio-rs/tracing/issues/73) in the tracing
repo which I've added our +1 to. It would provide the functionality we
are looking for if implemented.