add a mitigation to help with studio reporting
Users report that studio reporting fails after some time (days?). The
problem has been isolated to the `add` function in ReporterClient. The
root cause of that fail is hard to identify, so until a root causes is
found, this mitigation attempts to workaround the problem.

If an `add` call failes, the router will now create a new ReporterClient
and immediately try to `add` the report again using the new client.

A log warning is also added to report this event.