go through spaceport interactions and try to clean them up
The main fix here is that we shutdown the spaceport for the studio
reporting if we got a schema reload.

This change removes the shutdown spaceport on telemetry functionality
and instead relies on the fact that when a receiver gets no more traffic
it will terminate. This means we don't need to actively shut down a
spaceport connection.

There are some extra changes in here which I implemented along the way
to figuring out the main cause:
 - reconnect for recycled connections (may not be needed)
 - extra debug logging to figure out issues