improve studio reporting robustness (#1928)
fixes: #1903 

The main fix here is that we previously shutdown the spaceport for the
studio
reporting if we got a schema or config reload. That meant studio
reporting would
never resume until the router re-started.
    
This change removes the shutdown spaceport on telemetry functionality
and instead relies on the fact that when a receiver gets no more traffic
it will terminate. This means we don't need to actively shut down a
spaceport connection.
    
There are some extra changes in here which I implemented along the way
to figuring out the main cause:
- reconnect for recycled connections (not strictly required, but
improves connection robustness)
  - extra logging to clarify control flow

***
fixes: #1893 

Rather than allowing deadpool to choose the maximum size of a
connection pool based on the execution environment, fix this at a
reasonable size (128) for all execution environments.