replace lrucache with hashmap
I was searching for memory issues in the router and noted that we used
lru cache where we probably only needed a hashmap.

The cache is loaded and emptied over a fairly tight iteration in the
export function, so I don't know why this is using a long lived cache. I
also noted that the API complained about the fact it had to do a double
lookup, which isn't required with hashmap.

In my testing I never found more than 10 entries in the map, so I've
configured it to accept 64.