replace hotwatch with notify (#1964)
fixes: #1476  
fixes: #1695

The change replaces `hotwatch` with `notify` and makes use of the
"recommended" watcher feature which attempts to pick a file watcher
which is best suited to the execution environment. I've also improved
the robustness of the test and the way in which events are conveyed from
the file watcher to the consumer (retries on full to prevent lost event
notifications).

I've tested this manually on my local system (Macos M1) and confirmed:
 - test suite works
 - docker with --hot-reload works
 - kubernetes in k3d works*    

* Note: [Limitations in kubernetes configmap
handling](https://kubernetes.io/docs/concepts/configuration/configmap/#mounted-configmaps-are-updated-automatically)
mean that router pods in kubernetes won't react to changes to configmaps
mounted to containers which specify a subPath (which the router helm
chart currently does). Changes to schema will trigger a hot-reload.

As a drive-by: I've made a slight modification to the helm chart which
removes the use of subPath. If we don't specify the subPath and mount
the chart at /app/ we get the same result but with the added benefit of
configMap updates by k8s.