Faster scaffold test: share `target` and `Cargo.lock` with main workspace
Since this test is typically run together with other tests,
the apollo-router and its dependencies will be already built.
This sharing avoids recompiling some of the dependencies,
but unfortunately not all because of some differences in feature selection.

For local development however, repeated runs of the test will be much faster
since the target directory is preserved across runs.

RUSTFLAGS is replaced with a crate attributes so that it doesn’t affected
dependencies (which then don’t all need to be recompiled)