tweaked approach to dedup management
There is a lot of hackery in here which still needs to be cleaned up:
 - proper error handling
 - proper usage reporting

However, as a test of the deduplicate crate, it is good. The crate
preserves the fast behaviour of the existing approach and also doesn't
lock up (hang) which is the drawback with the existing approach.

I've tested it fairly extensively on my laptop, but it will need more
rigorous testing yet. I expect to be able to complete that over the next
couple of days.

Once more, deduplication and caching is a solved problem! :)