add trace_id in logs to identify all logs related to a request/error (#1982)
close #1981 

It automatically adds a `trace_id` on logs to identify which log is
related to a specific request. Also adds `trace_id` to an example error
to have an example.

This PR also refactor the way we handled formatters before, this
refactor will let us be more flexible on which specific fields we want
to display in logs. For now it keeps the current behavior which do not
display any fields coming from current span and parent spans, only
fields written directly in the log macro.

Example of an error response:
```json
{
  "data": {},
  "errors": [
    {
      "message": "HTTP fetch failed from 'accounts': HTTP fetch failed from 'accounts': error trying to connect: tcp connect error: Connection refused (os error 111)",
      "path": [],
      "extensions": {
        "type": "SubrequestHttpError",
        "service": "accounts",
        "reason": "HTTP fetch failed from 'accounts': error trying to connect: tcp connect error: Connection refused (os error 111)"
      }
    }
  ]
}
```

with response header:
```
apollo_trace_id: 5e6a6bda8d0dca26e5aec14dafa6d96f
```

with related logs:

```logs
2022-10-21T15:17:45.562553Z ERROR [trace_id=5e6a6bda8d0dca26e5aec14dafa6d96f] apollo_router::services::subgraph_service: fetch_error="hyper::Error(Connect, ConnectError(\"tcp connect error\", Os { code: 111, kind: ConnectionRefused, message: \"Connection refused\" }))"
2022-10-21T15:17:45.565768Z ERROR [trace_id=5e6a6bda8d0dca26e5aec14dafa6d96f] apollo_router::query_planner::execution: Fetch error: HTTP fetch failed from 'accounts': HTTP fetch failed from 'accounts': error trying to connect: tcp connect error: Connection refused (os error 111)
```

Signed-off-by: Benjamin Coenen <5719034+bnjjj@users.noreply.github.com>