improve the deduplication logic (#2014)
Close examination revealed that some judicious tweaking of the locking
and spawning would improve the behaviour.

In performance testing this performs the same as current (dev) or using
the deduplicate crate. I can't reproduce the hanging problem with dev
with this set of changes.

The rationale for the changes are as follows:

- We need to spawn the sentinel **before** we insert the waiter entry or
failure after inserting, but before spawning a sentinel, will cause
issues.
- Notification of waiters from the "first" entry **must be
synchronised** under the "wait_map" lock, or else that becomes racy and
subscribers may be missed. That notification must be immediately
preceded by the removal of the wait_map entry.

fixes: #1984