allow router installer to install aarch64 binaries
In the past, we didn't have aarch64 binaries. We hacked our install
script so that we downloaded x86_64 binaries and relied on magic (aka
chip emulation) to run the binary.

We build aarch64 binaries now, so we need to update the installer to
install them and give a performance boost to our linux on aarch64
user base.