allow router installer to install aarch64 binaries (#2033)
In the past, we didn't have aarch64 binaries. We hacked our install
script so that we downloaded x86_64 binaries and relied on magic (aka
chip emulation) to run the binary.

We build aarch64 binaries now, so we need to update the installer to
install them and give a performance boost to our linux on aarch64 user
base.

I manually tested the change by running a debian docker image on my M1
laptop and using the modified script to install a router into it.