apply tower best practices to services when cloning inner services (#2030)
When I was investigating connection resets I noted that we weren't
following tower documentation with respect to [cloning inner
services](https://docs.rs/tower/latest/tower/trait.Service.html#be-careful-when-cloning-inner-services).

This PR addresses that. The change to the supergraph seems to be most
invasive. I have run the tests manually many times and the change seems
good, but probably worth closer examination.