replace the BatchSpanProcessor with SimpleSpanProcessor
rust-opentelemetry's BatchSpanProcessor does not handle well larger
loads, even with sampling: some traces are not sent, some are sent but
miss some spans.  Benchmarking has shown that with the
SimpleSpanProcessor, all the traces are sent with all their spans, at
the cost of some latency and CPU time, but that is easily fixed by
sampling.

This also removes the filter span processor, because
SimpleSpanProcessor::new is not public, so we cannot create it and apply
the filter on it. It is replaces with a span exporter filter, that we
can apply directly over the filter