apply traffic shaping on subgraph directly (#2034)
The plugin infrastructure works on `BoxService` instances, and makes no
guarantee on plugin ordering.
The traffic shaping plugin needs a clonable inner service, and should
run right before calling
the underlying service. So this changes the traffic plugin application
so it can work directly
on the underlying service. It is still a plugin though, so it keeps the
same configuration.