Fix helm values command (#2088)
This is a tiny PR but I think it's a valid fix. 

When I tried to run `helm show values apollographql/router` as suggested
in the README I got the error `Error: repo apollographql not found`.

I believe the command should actually be `helm show values
oci://ghcr.io/apollographql/helm-charts/router` which works as expected
for me.