Wait for current tracer provider to shutdown when dropping telemetry (#2191)
We spawn a detached thread currently. That's probably not a good idea
since the router may terminate before the thread terminates and that
will likely cause spans to be lost.

Let's wait for the shutdown for up to 5 seconds in the `drop` and then
print a warning message if there is an issue.

The figure of 5 seconds is not configurable, but could be made so in
future if required.