first push to gh with some working functionality
Enough code to be able to externalise a request and process the data we
receive from our co-processor.

Many bits need to be addressed:
 - testing
 - responses
 - deciding if simple config is good.

The current (simple) config I'm working with looks like this:

```yaml
plugins:
  experimental.external:
    url: http://127.0.0.1:8081
    request:
      headers: true
      context: true
      body: true
      sdl: true
    response:
      headers: true
```

The various fields: headers -> sdl are all optional and the default is
false if not specified. In the above I'm sending all the data on the
request and I'm sending just headers on the response.

We should maybe add which step/steps(?) this callout is made from at
some point. For now, it's hard coded to be just the RouterService.