replace http status block with control block
Feedback from lenny and team was that control was preferred to http
status codes with implicit control.

Now possible to return a "control" instruction from co-processor. That
can either be "Continue" or "Break" with appropriate status code. If
break, the body is assumed to be the content of the break message. If
the code is a 2xx, then the body is assumed to contain "data", otherwise
it is treated as an error message.

This may need tidying up at some point, but that's what it does for now.