reconstruct deferred queries with knowledge about fragments (#2109)
Fix #2105 
Fix #2099 

When we are using `@defer`, response formatting must apply on a subset
of the query (primary or deferred), that is reconstructed from
information provided by the query planner: a path into the response and
a subselection. Previously, the query planner did not add fragments to
that path, which resulted in reconstruction issues for queries like:

```graphql
query {
  me {
    ... on User {
      id
      fullName
      memberships {
        permission
        account {
          ... on Account @defer {
            name
          }
        }
      }
    }
  }
}
```

we would get a deferred part with the path `/me/memberships/account` and
the subselection `{ name }`. The path does not give enough information
to reconstruct the query when `me` returns an interface: we cannot know
which type is used.

The query planner 2.2.2 introduces a fragment in the path to solve that,
like `/me/... on User/memberships/account`. It also removes the
'flatten' element from deferred nodes paths, so the code has to be
adapted a bit.