replace notify recommended watcher with poll watcher (#2276)
To get consistent behaviour across multiple platforms and file systems.

fixes: #2245
  
The watcher may not be as reactive, but I think it's reactive enough for
things which shouldn't change frequently.

As a bonus, we no longer get events if the content is the same, so less
spurious reloads, which is good. That means I had to modify the tests
though to use different contents.

The rhai file watching appeared to be working fine, but to be safe I
converted it to the PollWatcher as well.