move APQ and EnsureQueryPresence in the router service
When the router service was created, the APQ functionality that was
performed at the axum level, before calling the supergraph service, was
moved to a layer above the supergraph service. Since it requires
AsyncCheckPoint (to asynchronously call a database to get the query), it
required the inner service to be cloned, so we had to reintroduce a
Buffer layer.
There was also a semantic issue here: we assume that a supergraph
request contains a valid graphql request, but instead, when the router
service creates it, it lets layers in the supergraph service deal with
APQ and checking that the query is present.

This refactrs the APQ and EnsureQueryPresence layers, to move them in
the router service's code, after the supergraph request is created, but
before it is passed to the supergraph service. This reduces the code
drastically (a few branches instead of two entire layers). The tests
from EnsureQueryPresence are now moved to the router service.