warm up the query planner cache with previously cached plans
On schema updates, the query plan cache must be rebuilt from scratch,
which causes latency instabilities. This adds a warm up phase, that gets
the queries already in cache for the previous schema, and runs them
through the query planner with the new schema and caches them, before
switching the router to use the new schema. The warmed up queries will
use the cache directly and keep the same latency. This can make schema
updates slow and CPU intensive though, as we have to wait for all the
query plans to be generated.