docs: Retroactively update `federation-version-support.mdx` for v1.6.0.
This should have been updated as part of the release process for the 1.6.0
release to indicate that Federation v2.2.2 was the active version as of that
release.  Apologies to the folks who rely on this information being accurate
to do their job correctly.

In the future, we will introduce automation to eliminate the human factor of
this, and that work is being tracked in https://github.com/apollographql/router/issues/2261.