Remove destructive `git reset --hard` command which destroyed my local changes
There should be no need to use something as agressive as `git reset --hard`
in a script.  If such behavior is required, it should be initiated directly
by the user, never in an automated fashion since it's impossible to guarantee
that the workspace has no local changes which would be sacrificed to the Git
gods.