Require a pristine Git checkout of known files prior to releasing
We want to avoid cases where changes which are staged/gathered locally
somehow impact the software that gets released - for example, allowing a
local change to inadvertently get checked into the release commit.

Prior to initiating the release, we will now assert that the checkout
doesn't have any changes to "known files" (i.e., files which are committed
to Git).

Currently, this is configured to IGNORE untracked files when making this
determination, though we could also choose a more aggressive option that
didn't relax the assessment against untracked files.  Practically speaking,
an untracked file shouldn't impact the release since we have a CI pipeline
responsible for the releasing. With this considered, it seemed like a
reasonable trade-off since many on the team have many untracked files in
their local checkouts that they use for various purposes (e.g, Configuration
files, etc.) which would otherwise need to be fully stashed prior to releasing.

In my experience, this change also made it a bit more tolerable to iterate
on the release xtask script during development since it was a force-function
that made sure I reset my changes before doing it again.  The DOWNSIDE is
that it becomes necessary to disable the `ensure_pristine_checkout` call,
but that seems like an okay trade-off.